==> Instacart Basket Analysis - Tynan Englert /02 Scripts/01 Subsetting.ipynb <==
# 4.4 Subsetting Data

# Importing Libararies

#importing pandas, numpy and os 
import pandas as pd
import numpy as np
import os 

# Importing Data 

# importing orders data set 
path = r'/Users/brookecoffey/10-07-2023 Instacart Basket Analysis'

df_ords = pd.read_csv(os.path.join(path, '02 Data ', 'Original Data ', 'orders.csv'), index_col = False)

# importing products data set 
df_prods = pd.read_csv(os.path.join(path, '02 Data ', 'Original Data ', 'products.csv'), index_col = False)

# Transforming Data Examples 

# Dropping eval column from orders dataframe
df_ords.drop(columns = ['eval_set'])

# Count values for a certain column
df_ords['days_since_prior_order'].value_counts(dropna = False)

# Changing Column Names 
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# Use head function to check if the renaming was correct 
df_ords.head()

# changing data type of order_id column 
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

# Importing departments Data set 

df_dep = pd.read_csv(os.path.join(path, '02 Data ', 'Original Data ', 'departments.csv'), index_col = False)

df_dep.head()

# Transposing from wide format 
df_dep.T

# Creating a new data frame from the above function 
df_dep_t = df_dep.T

df_dep_t

# Removing the 0 at the top of the df_dep_t
# STEP 1 is creating a new index
df_dep_t. reset_index()

#STEP 2 - Create a new header 
new_header = df_dep_t.iloc[0]

new_header

# STEP 3 - Remove first row from dataframe
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# STEP 4 - Add header to new dataframe 
df_dep_t_new.columns = new_header

df_dep_t_new

# Creating a data dictionary 
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

# Use data dictionary to look up name of department id 19
print(data_dict.get('19'))

# Step 1 of subsetting query, identifies cases on True/False basis of whether an item is in category id 19
df_prods['department_id']==19

# Step 2 of subsetting query, selects those rows that are 'True' for being in category id 19
df_prods[df_prods['department_id']==19]

# Step 3 of subsetting query is creating a new dataframe from that query 
df_snacks = df_prods[df_prods['department_id']==19]

df_snacks.head()

# Alternative to original subsetting query using loc function
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_2

# Other alternative to original subsetting query using loc and 'isin' - HELPFUL FOR LISTS
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks_3

# EXERCISE PROMPTS

### 01 Find another identifier variable in the df_ords dataframe that doesn’t need to be included in your analysis as a numeric variable and change it to a suitable format.

df_ords.head()

df_ords.dtypes

# Changing user_id to a string from an integer 
df_ords['user_id'] = df_ords['user_id'].astype(str)

# Checking that the change was correct and confirming new data type of user_id column
df_ords['user_id'].dtype

### 02 Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the data frame.

# Renaming the order_number column to more accurately reflect what the value represents 
df_ords.rename (columns = {'order_number' : 'order_number_by_customer'}, inplace = False)

### 03 Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.

# Count of frequencies for the order_hour_of_day Column
df_ords['order_hour_of_day'].value_counts(dropna = False)

### 03 Answer: Hour 10 is the busiest hour for placing orders

### 04 Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.

# Use data dictionary to look up name of department id 4
print(data_dict.get('4'))

### 05 The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.

# use data dictionary to determine what category breakfast foods are
data_dict

# create a subset for only breakfast foods 
df_breakfast = df_prods[df_prods['department_id']==4]

# confirm that the new dataframe consists only of breakfast items
df_breakfast

### 06 They’d also like to see details about products that customers might use to throw dinner parties. Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. You’ll need to present this subset to your client.

# Determine category id for the above departments using data dictionary (5, 20, 7, 12)
data_dict

# Create a dataframe consisting of products only in departments 5, 20, 7, 12 
df_party_supplies = df_prods.loc[df_prods['department_id'].isin([5, 20, 7, 12])]

# Double-check new dataframe
df_party_supplies

### 07 It’s important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?

### 07 A: 7650 Rows 

### 08 Someone from the data engineers team in Instacart thinks they’ve spotted something strange about the customer with a "user_id" of “1.” Extract all the information you can about this user.

# Create a dataframe with all information for user_id 1
df_user_a = df_ords.loc[df_ords['user_id']=='1']

# confirm that the dataframe is for user_id 1 specifically
df_user_a

### 09 You also need to provide some details about this user’s behavior. What basic stats can you provide based on the information you have?

# use describe function to provide descriptive stats for user A
df_user_a.describe()

# Exporting Dataframes

# exporting the orders dataframe to csv 
df_ords.to_csv(os.path.join(path, '02 Data ', 'Prepared Data ', 'orders_wrangled.csv'))

# exporting dep_t_new dataframe to csv 
df_dep_t_new.to_csv(os.path.join(path, '02 Data ', 'Prepared Data ', 'departments_wrangled.csv'))
