==> Instacart Basket Analysis - Tynan Englert /02 Scripts/02 Data Consistency Checks.ipynb <==
# Exercise 4.5 - Data Consistency Checks

# Importing Libraries and Datasets

# Import pandas, numpy and os 
import pandas as pd
import numpy as np 
import os

# import products.csv from Original Data folder and orders_wrangled.csv from Prepared Data Folder
path = r'/Users/brookecoffey/10-07-2023 Instacart Basket Analysis'

df_prods = pd.read_csv(os.path.join(path, '02 Data ', 'Original Data ', 'products.csv'), index_col = False)

df_ords = pd.read_csv(os.path.join(path, '02 Data ', 'Prepared Data ', 'orders_wrangled.csv'), index_col = False)

df_ords.describe()

# Mixed-Type Data

# Create a test dataframe
df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# Function to find any mixed type columns
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# Changing data type of mixed column
df_test['mix'] = df_test['mix'].astype('str')

# Missing Values

# Finding missing values in products table 
df_prods.isnull().sum()

# Create a subset of the false values for the product_name column
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# get shape of existing df_prods dataframe
df_prods.shape

# Create new dataframe excluding the products with null product name
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# Compare shape of new dataframe with old dataframe
df_prods_clean.shape

# Alternative method of dropping null values - this overwrites the existing dataframe
# df_prods.dropna(inplace = True)
# df_prods.dropna(subset = [‘product_name’], inplace = True)

# Duplicates

# Looking for full duplicates (exact same values for multiple rows)
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

# checking current shape before deleting duplicates
df_prods_clean.shape

# Create new dataframe with no duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# Exercise Prompts

# Describe the df_ords dataframe 
df_ords.describe()

## A2: All columns look correct with what one would expect.  order_id, user_id, and order_number all have minimums of 1, rather than 0.  order_day_of_week, order_hour_of_day and days_since_prior_order all have minimums of 0, appropriate maximium values for those fields (i.e. no more than 7 days in a week or 24 hours in a day), and appropriate spreads (i.e. 50% of day of week should be 3, and 13 for hour of day).  

# Check for mixed-type data in your df_ords dataframe
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

## A3: Using the command above from the reading, substituting df_ords for df_test, I didn't find any mixed-type columns

# Run a check for missing values in your df_ords dataframe.
df_ords.isnull().sum()

## A4: The only column with missing values is days_since_prior_order.  This would most likely be due to first-time orders that would have a null value for that field, as there is no last order to use as a reference point.  If that is in fact true, then this column doesn't require any filtering or imputing.  

# Address the missing values using an appropriate method.
df_ords['first_order'] = np.where(df_ords['days_since_prior_order'].isnull() == True, True, False)

df_ords.head()

## A5: Becuase the null values in the days_since_prior_order column are valuable to preserve, I added another column "first_order" that specifies if the order was the first order of that customer.  If the days_since_prior_order column shows a Null value, the first_order column will return a boolean response, with True meaning that it is the first order and False meaning that it is not the first order.  This would kind of be a restatement of the order_number column.  

# Run a check for duplicate values in your df_ords data.
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

## A6: No duplicate records were found 

df_ords.shape

# Exporting Files 

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data ', 'Prepared Data ', 'products_checked.csv'))

df_ords.to_csv(os.path.join(path, '02 Data ', 'Prepared Data ', 'orders_checked.csv'))
