==> Instacart Basket Analysis - Tynan Englert /02 Scripts/03 Deriving New Variables.ipynb <==
# 4.7 - Deriving New Variables 

## 01: Importing Libraries

# Import numpy, pandas, os
import pandas as pd
import numpy as np 
import os

## 02: Importing Data

# Set path for data 
path = r'/Users/brookecoffey/10-07-2023 Instacart Basket Analysis'

# Import ords_prods_merged dataset
ords_prods_merged = pd.read_pickle(os.path.join(path, '02 Data ', 'Prepared Data ', 'orders_products_merged.pkl'))

# Double-check shape before beginning 
ords_prods_merged.shape

# Create a subset of only the first million rows to avoid data issues 
df = ords_prods_merged[:1000000]

df.shape

## 03 If-Statements with User-Defined Functions

# defining price range arguments 

def price_label(row):
    if row['prices'] <= 5: 
        return 'Low-range Product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid-range Product'
    elif row['prices'] > 15:
        return 'High range'
    else: return 'Not enough data'

# applying new arguments to the dataframe 
df['price_range'] = df.apply(price_label, axis=1)

# double check output 
df.shape

# check frequency of price_range column in subset
df['price_range'].value_counts()

# find max price present in the subset 
df['prices'].max()

## 03 If-Statements with loc() function 

# create high range category using loc() function 
ords_prods_merged.loc[ords_prods_merged['prices'] > 15, 'price_range'] = 'High-range product'

# create mid-range category using loc() function
ords_prods_merged.loc[(ords_prods_merged['prices'] <= 15) & (ords_prods_merged['prices'] > 5), 'price_range'] = 'Mid-range product' 


# create low-range category using loc() function
ords_prods_merged.loc[ords_prods_merged['prices'] <= 5, 'price_range'] = 'Low-range product'

# count output of price ranges 
ords_prods_merged['price_range'].value_counts()

ords_prods_merged.shape

# apply same price range labels to subset 
df.loc[df['prices'] > 15, 'price_range'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range'] = 'Mid-range Product'

df.loc[df['prices'] <= 5, 'price_range'] = 'Low-range Product'

df['price_range'].value_counts()

df.shape

## 03 If Statements with For loops 

# using a for loop, calculate how busy each day of the week is.  
# Step 1, calculate frequency for each day of the week 
ords_prods_merged['order_dow'].value_counts(dropna = False)

# Creating for loop to create a new column for how busy each day is 
result = []

for value in ords_prods_merged["order_dow"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

# add result column to the dataframe
ords_prods_merged['busiest_day'] = result

# compare frequency against original count to confirm 
ords_prods_merged['busiest_day'].value_counts(dropna = False)

# 04 Exercise 4.7 Prompts

# double check ords_prods_merged df
ords_prods_merged.head()

### 01 Suppose your clients have changed their minds about the labels you created in your “busiest_day” column. Now, they want “Busiest day” to become “Busiest days” (plural). This label should correspond with the two busiest days of the week as opposed to the single busiest day. At the same time, they’d also like to know the two slowest days. Create a new column for this using a suitable method.

# using a for loop, calculate how busy each day of the week is.  
# Step 1, calculate frequency for each day of the week 
ords_prods_merged['order_dow'].value_counts(dropna = False)

# days 0 and 1 are the two busiest days, while days 3 and 4 are the least busy days 
# Creating for loop to create a new column for how busy each day is 
result_2 = []

for value in ords_prods_merged["order_dow"]:
  if value == 1 or value == 0:
    result_2.append("Busiest day")
  elif value == 3 or value == 4:
    result_2.append("Least busy")
  else:
    result_2.append("Regularly busy")


result_2

# add result column to the dataframe
ords_prods_merged['busiest_days'] = result_2

# Check the values of this new column for accuracy. Note any observations in markdown format.
ords_prods_merged['busiest_days'].value_counts(dropna = False)

ords_prods_merged['busiest_day'].value_counts(dropna = False)

#### Observations: The new "busiest day" category is the sum of days 0 and 1, the "least busy" category is the sum of days 3 and 4, and the "regularly busy" category is the sum of the remaining days.  This confirms that the column is calculating the totals correctly

### 02 When too many users make Instacart orders at the same time, the app freezes. The senior technical officer at Instacart wants you to identify the busiest hours of the day. Rather than by hour, they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.” Create a new column containing these labels called “busiest_period_of_day.”

# Check frequency of values for order_hour_of_day column 
ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

# divide the above output into three equal groups to fit categories above
result_3 = []

for value in ords_prods_merged["order_hour_of_day"]:
  if value in [10, 11, 14, 15, 13, 12, 16, 9] :
    result_3.append("Most orders")
  elif value in [17, 8, 18, 19, 20, 7, 21, 22]:
    result_3.append("Fewest orders")
  else:
    result_3.append("Average orders")

# add column to the final table 
ords_prods_merged['busiest_period_of_day'] = result_3

# print frequemcy for the new column
ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

# see basic output of full dataframe
ords_prods_merged.head()

# 05 Exporting Data

# exporting ords_prods_merged into pickle 
ords_prods_merged.to_pickle(os.path.join(path, '02 Data ', 'Prepared Data ', 'ords_prods_merged_updated.pkl'))
