==> Instacart Basket Analysis - Tynan Englert /02 Scripts/04 Grouping Data and Aggregating Variables.ipynb <==
# Exercise 4.8 Grouping Data and Aggregating Variables

## Importing Libraries

import numpy as np 
import pandas as pd
import os

## Importing Data

# Set path for data 
path = r'/Users/brookecoffey/10-07-2023 Instacart Basket Analysis'

# Import ords_prods_merged dataset
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merged_updated.pkl'))

## 02/03 In this Exercise, you learned how to find the aggregated mean of the “order_number” column grouped by “department_id” for a subset of your dataframe. Now, repeat this process for the entire dataframe.

# Create a subset for comparison 
df = ords_prods_merge[:1000000]

df.shape

# Group whole dataframe by department_id and find aggregated mean 
ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

# do the same aggregate function above for the subset for comparison
df.groupby('department_id').agg({'order_number': ['mean']})

### The aggregate calculation for the whole dataset contains all of the department id's rather than a small slice.  Departments that were included in the slice have slightly different aggregate means in the whole dataframe output because additional columns are taken into consideration that were excluded from the sliced dataframe

## 04 Follow the instructions in the Exercise for creating a loyalty flag for existing customers using the transform() and loc() functions.


# Create the max order column for each user id
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

# Create the sub-categories using the max order column 
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular Customer'
ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Check output
ords_prods_merge.head()

## 05 The marketing team at Instacart wants to know whether there’s a difference between the spending habits of the three types of customers you identified. Use the loyalty flag you created and check the basic statistics of the product prices for each loyalty category (Loyal Customer, Regular Customer, and New Customer). What you’re trying to determine is whether the prices of products purchased by loyal customers differ from those purchased by regular or new customers.

# Group dataframe by loyalty group
# For each loyalty group, calculate summary statistics (mean, min, max)
ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max']})

## 06 The team now wants to target different types of spenders in their marketing campaigns. This can be achieved by looking at the prices of the items people are buying. Create a spending flag for each user based on the average price across all their orders using the following criteria:

## If the mean of the prices of products purchased by a user is lower than 10, then flag them as a “Low spender.” 
## If the mean of the prices of products purchased by a user is higher than or equal to 10, then flag them as a “High spender.”

# Create a column to find the average price of purchases grouped by user id 
ords_prods_merge['avg_order_price'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

# Create the sub-categories using the avg order price column 
ords_prods_merge.loc[ords_prods_merge['avg_order_price'] >= 10, 'spending_flag'] = 'High Spender'
ords_prods_merge.loc[ords_prods_merge['avg_order_price'] < 10, 'spending_flag'] = 'Low Spender'

ords_prods_merge['spending_flag'].value_counts(dropna = False)

# Check output 
ords_prods_merge.head()

## 07 In order to send relevant notifications to users within the app (for instance, asking users if they want to buy the same item again), the Instacart team wants you to determine frequent versus non-frequent customers. Create an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column. The criteria for the flag should be as follows:

## If the median of “days_since_prior_order” is higher than 20, then the customer should be labeled a “Non-frequent customer.”

## If the median is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”

## If the median is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”

# Creating a column to find the median days since prior order for each user id 
ords_prods_merge['median_days_since_prior_order'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

# Create the sub-categories using the median days since prior order column created above
ords_prods_merge.loc[ords_prods_merge['median_days_since_prior_order'] > 20, 'frequency_tag'] = 'Non-frequent Customer'
ords_prods_merge.loc[(ords_prods_merge['median_days_since_prior_order'] > 10) & (ords_prods_merge['median_days_since_prior_order'] <= 20), 'frequency_tag'] = 'Regular Customer'
ords_prods_merge.loc[ords_prods_merge['median_days_since_prior_order'] <= 10, 'frequency_tag'] = 'Frequent Customer'

# Check output 
ords_prods_merge['frequency_tag'].value_counts(dropna = False)

## 08 Exporting dataframe 

ords_prods_merge.to_pickle(os.path.join(path, '02 Data ', 'Prepared Data ', 'ords_prods_merge_updated2.pkl'))
